% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{chooseLoc}
\alias{chooseLoc}
\title{chooseLoc}
\usage{
chooseLoc(hab.prob, pd.rate, steps, lambda, coef.d, coef.r, blank.rast,
  matsize, ...)
}
\arguments{
\item{hab.prob}{matrix, the habitat matrix with cell values representing the
habitat-only probabilities (uncorrected for the mixing proportions with
distance). Typically the second element of the list returned by
simlandscapes.}

\item{pd.rate}{numeric, the probability of returning to the central place to
"deliver prey" to nest/den/etc. on any given step (given the agent isnot
currently located at the central place.) Must be <1.}

\item{steps}{integer, the number of steps the agent should take}

\item{lambda}{numeric, the strength of the declining exponental function that controls
how "reluctant" the agent is to move large distances. Must be <1.}

\item{coef.d}{numeric, can be factor or proportion representing the relative
strength of the preference to move short steps compared to the strength of
the preference for habitat.}

\item{coef.r}{numeric, can be factor or proportion representing the relative
strength of the preference for habitat compared to the strength of
the preference for short steps.}

\item{matsize}{integer, the size in number of cells of one side of the square
matrix on which the simulation is being run.}

\item{...}{}
}
\value{
Returns the list of positions for the simulated individual.
}
\description{
A function to simulate a single individual moving with varying mixtures of
preference for habitat and/or short steps.
}
