% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{compilerangesSDM}
\alias{compilerangesSDM}
\title{compilerangesSDM}
\usage{
compilerangesSDM(nulldata, habdata, condata, A.coef, radius)
}
\arguments{
\item{nulldata}{data object output by `repRand`.}

\item{habdata}{data object output by `repHab`.}

\item{A.coef}{numeric, set of habitat preference coefficients supplied to
`repHab`.}

\item{radius}{numeric, set of conspecific attraction radii supplied to
`repCon`.}

\item{condatat}{data object output by `repCon`.}
}
\value{
a dataframe containg the sampling distribution ranges for all
simulation models supplied.
}
\description{
compilerangesSDM
}
