% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{convert.cell}
\alias{convert.cell}
\title{convert.cell}
\usage{
convert.cell(mat, ...)
}
\arguments{
\item{mat}{Matrix to be converted (typically used to convert the output of pref.strength)}

\item{...}{Additional arguments as necessary.}
}
\value{
A matrix with the values of each cell converted to relative probabilities,
 normalized byt the sum of the values of the matrix
}
\description{
Converts the values of each cell in a matrix
to probabilities by normalizing each cell's value by the sum of the entire matrix.
}
