% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{create.land}
\alias{create.land}
\title{create.land}
\usage{
create.land(n.clusters, size.clusters, x.mat, count.max = 200, ...)
}
\arguments{
\item{n.clusters}{A number. Target number of clusters to grow on the landscape}

\item{size.clusters}{A number. Target size (measured in number cells) of each patch.}

\item{x.mat}{A matrix whose cell values all = 1}

\item{count.max}{A number. Maximum number of iterations through the growing process (default is 200).}

\item{...}{Additional arguments as necessary.}
}
\value{
A matrix with each cell that is a member of a patch identified by it's patch ID. Cells not inculded in a patch are listed as NA
}
\description{
Creates a simulated patchy habitat matrix based on user defined inputs.
Uses the expand function to grow each habitat patch.
}
