% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{createKDE}
\alias{createKDE}
\title{createKDE}
\usage{
createKDE(mov)
}
\arguments{
\item{mov}{movement data supplied in format produced by `chooseLoc`}
}
\description{
Ceeates a kernel density estimate based on simulated movement data
}
