% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eatfood.R
\name{eatfoodonce}
\alias{eatfoodonce}
\title{eatfoodonce}
\usage{
eatfoodonce(num_sources, diet_prop, food)
}
\arguments{
\item{num_sources}{integer, the number of food sources to choose from.
Can not exceed the actual number of available food sources.}

\item{diet_prop}{vector, the relative probability of selecting a source.
Must have an entry for each source.}

\item{food}{list, the simulated food sources from which to select.}
}
\value{
A vector containing the carbon, then nitrogen values for the prey
item selected.
}
\description{
Simulates a single consumption event by a consumer and returns the carbon and
nitrogen values for the consumed prey.  Meant to be call by `eatfoodsteps`
and `eatfood`.
}
