% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eatfood.R
\name{eatfoodsteps}
\alias{eatfoodsteps}
\title{eatfoodsteps}
\usage{
eatfoodsteps(steps, num_sources, diet_prop, food)
}
\arguments{
\item{steps}{Integer, the number of time steps to simulate}

\item{num_sources}{integer, the number of food sources to choose from.
Can not exceed the actual number of available food sources. Passed to
`eatfoodonce`.}

\item{diet_prop}{vector, the relative probability of selecting a source.
Must have an entry for each source. Passed to `eatfoodonce`.}

\item{food}{list, the simulated food sources from which to select. Passed to `eatfoodonce`.}
}
\value{
A list of consumer-sampled prey isotope values. Each element of the
list is a step containing the C and N values for the prey item sampled.
}
\description{
Wrapper function to interate `eatfoodonce` across multiple time steps (for
a sinlge individual). Returns a matrix of C and N values sampled by the
individual over `steps` number of steps.
}
