% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{expand}
\alias{expand}
\title{expand}
\usage{
expand(x, n.size, start)
}
\arguments{
\item{x}{A matrix whose cell values all = 1}

\item{n.size}{A number representing the target size for each cluster}

\item{start}{A number between 1 and the number of cells in matrx x indicating the algorithm starting position.}
}
\value{
A a grid of generated numbers of approximately n.size
}
\description{
cellular automata-like function to "grow" habitat patches
}
\details{
Function to expand a patch randomly within indicator array x (wherein
1=unoccupied) by n.size cells, beginning at index start.
Algorithm based on and modified from:
https://gis.stackexchange.com/questions/60562/creating-randomly-shaped-clumps-of-
cells-in-a-raster-from-seeds-of-1-cell-pixel
}
