% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettissueiso.R
\name{formatiso}
\alias{formatiso}
\title{formatiso}
\usage{
formatiso(iso, filename, writefile = T, returnobject = F)
}
\arguments{
\item{iso}{List of simulated "sampled" C and N values for each simulated
consumer. Note this should be the output from `getiso`.}

\item{filename}{String, the name of the file to be saved as a .csv.}

\item{writefile}{Boolean, whether to write the .csv to file. Default is T.}

\item{returnobject}{Boolean, whether to return a dataframe object. Default
is F}
}
\value{
A .csv written to file and/or a dataframe object.
}
\description{
Convenience fucntion to format simulated consumer tissue isotope values into
a dataframe with columns for d13C and d15N and each row being an indivdiual
"observation". Default is to write the object to a .csv file formatted for
use by `MixSIAR` for mixture decomposition analysis, but can also return a
dataframe as an object.
}
