% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{getcombos}
\alias{getcombos}
\title{getcombos}
\usage{
getcombos(X, Y)
}
\arguments{
\item{X}{The first list of model objects. It is suggested that each list
element be named.}

\item{Y}{The second list.}
}
\value{
A two element list which contains the paired combinations of the
supplied model lists
}
\description{
Creates 2 lists which, paired, provide each unidirectional pairwise
combination (outer product) of the two lists. Meant as a component function to
assist in creating heatmaps.
}
