% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{gethigh}
\alias{gethigh}
\title{gethigh}
\usage{
gethigh(range)
}
\arguments{
\item{range}{object output by getrangeSDM}
}
\description{
Convenience function to extract high values from `getrangeSDM` output.
}
