% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{getHRUsed}
\alias{getHRUsed}
\title{getHRUsed}
\usage{
getHRUsed(poly, hab.mat, matsize)
}
\arguments{
\item{poly}{A polygon of the format `SpatialPolygons` or
`SpatialPolygonsDataFrame`. Typically the output of the `getverticesHR`
function in the `adehabitatHR` package.}

\item{hab.mat}{The habitat classification matrix (first element of the list
generated by `simlandscapes`)}

\item{matsize}{The size of one side of the square matrix on which the
simualtion is run.}
}
\value{

}
\description{
Extracts the habitat classification of habitat matrix cells contained within
the boundaries of a supplied polygon (intended to be the isopleth of a
KDE-based home range)
}
