% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettissueiso.R
\name{getiso}
\alias{getiso}
\title{getiso}
\usage{
getiso(window, time, food)
}
\arguments{
\item{window}{Integer, the consumer isotope integration window representing
the number of steps across which prey isotope values are integrated in an
individual consumer to generate consumer tissue isotope values. Window must
be less than or equal to `time`. Passed to `gettissue`.}

\item{time}{Integer, the step at which the simulated consumer tissue sample
is to be taken. Must be less than or equal to the total number of steps
simulated. Passed to `gettissue`.}

\item{food}{The list of simulated prey consumption histories for all
individual consumers. Note this is the list generated by `eatfood`.}
}
\value{
A list of simulated "sampled" C and N values for each simulated
consumer. Each element of the list represents the sampled value for a single
individual consumer and in each list element is a vector of the simulated
tissue C and N values (in that order) produced from `gettissue`.
}
\description{
Wrapper function to iterate `gettissue` across the set of simulated consumers.
}
