% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{getlow}
\alias{getlow}
\title{getlow}
\usage{
getlow(range)
}
\arguments{
\item{range}{object output by getrangeSDM}
}
\value{
a single value for the low range estimate for a component of the
output of getrangeSDM
}
\description{
Convenience function to extract low values from `getrangeSDM` output.
}
