% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{getPropUsed}
\alias{getPropUsed}
\title{getPropUsed}
\usage{
getPropUsed(movelist, hab.mat, A.coef, matsize)
}
\arguments{
\item{movelist}{The list of locations visitied by agent in a single model
iteration.}

\item{hab.mat}{The habitat matrix on which the simulation was run, needs
version with 1's and `A.coef` (not probabilities of cell IDs)}

\item{A.coef}{`A.coef` used to simulate habitat - used as key to `hab.mat`}

\item{matsize}{The size of one side of the square matrix on which the
simualtion is run}
}
\description{
Function to get the proportion of used locations (from movement model) within
 the habitat matching the value of `A.coef`
}
