% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{getrangeSDM}
\alias{getrangeSDM}
\title{getrangeSDM}
\usage{
getrangeSDM(dat, ...)
}
\arguments{
\item{dat}{Dataset from which to extract range}

\item{...}{}
}
\value{
Returns a list containing the minimum and maximum values for the
input distribution.
}
\description{
Function to extract distribution range from simulated SDM output (i.e., the
list output by `repRand`, `repHab`, and/or `repCon`. Expects data of the SDM
simulation format and reports the minimum and maximum observed preference for
"Habitat A".
}
