% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettissueiso.R
\name{gettissue}
\alias{gettissue}
\title{gettissue}
\usage{
gettissue(foodlist, window, time)
}
\arguments{
\item{foodlist}{The list of food items sampled by an individual consumer.
This is the list generated by `eatfoodsteps`. Note that this function is
intededed to be called from `getiso` which iterates this function across
all individuals in a population of consumers (i.e., the list generated by
`eatfood`).}

\item{window}{Integer, the consumer isotope integration window representing
the number of steps across which prey isotope values are integrated in an
individual consumer to generate consumer tissue isotope values. Window must
be less than or equal to `time`.}

\item{time}{Integer, the step at which the simulated consumer tissue sample
is to be taken. Must be less than or equal to the total number of steps
simulated.}
}
\value{
A vector containing the C and N values for the simulated consumer
tissue.  Tissue values represent the mean isotope values for the prey
consumed between `time` - `window` and `time`.
}
\description{
Function to calculate tissue isotope values from a single simulated consumer.
}
