% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{makeDistProb}
\alias{makeDistProb}
\title{makeDistProb}
\usage{
makeDistProb(matsize, position, lambda)
}
\arguments{
\item{matsize}{numeric, size of one side of a square matrix on which to
perform calculations}

\item{position}{numeric, position from which to generate declining
probability}

\item{lambda}{numeric, strength of exponential decline ($\lambda$) where:
$x = \lambda^{-\lambda x}$}
}
\description{
Function to make a matrix of size `matsize` containing probabilities that
decline exponentially by distance from a given cell (`position`). Strength of
exponential decline is given by `lambda`
}
