% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefood.R
\name{makefood}
\alias{makefood}
\title{makefood}
\usage{
makefood(num_sources, popsize, mu_carb, sd_carb, mu_nit, sd_nit)
}
\arguments{
\item{num_sources}{integer, number of potential food sources to simulate}

\item{popsize}{vector whose length = 'num_sources', each entry of which is
the size of the diet source population to simulate}

\item{mu_carb}{vector whose length = 'num_sources', each entry of which is
the mean 13C value of the diet source population to simulate}

\item{sd_carb}{vector whose length = 'num_sources', each entry of which is
the standard deviation of the 13C of the diet source population to simulate
#' @param mu_nit vector whose length = 'num_sources', each entry of which is
the mean 15N value of the diet source population to simulate}

\item{sd_nit}{vector whose length = 'num_sources', each entry of which is
the standard deviation of the 15N of the diet source population to simulate}
}
\value{
A list, each element of which is a vector of isotope vales for each
simulated diet endpoint individual
}
\description{
Simulates the 13C isotope values for an arbitray number of potential diet
endpoints. Each endpoint population is simulated as a normally distributed
variable
}
