% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{matrixCellFromXY}
\alias{matrixCellFromXY}
\title{matrixCellFromXY}
\usage{
matrixCellFromXY(pos, matsize)
}
\arguments{
\item{pos}{vector, a position in the format `c(X, Y)` i.e. `c(col, row)`.
Intended to accept a single position in `move.list`}

\item{matsize}{integer, the size of the square matrix on which the simulation
is being run}
}
\value{
integer, the cell ID indicated by the X, Y coordinates.
}
\description{
Gets the cell ID within a matrix using the position format (`c(X, Y)`)
suppplied by `move.list`.
}
