% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{matrixPythagoras}
\alias{matrixPythagoras}
\title{matrixPythagoras}
\usage{
matrixPythagoras(position1, position2, IDmat)
}
\arguments{
\item{position1}{first positions in distance calc}

\item{position2}{second position in disttance calc}

\item{IDmat}{matrix, the ID matrix on which to calculate distance, must have
cell values matching 1-D indexing. See examples}
}
\value{
returns the pythogorean distance between two coordinates (supplied as
matirx indices)
}
\description{
Uses the Pythagorean theorem to calculate the strightline distance between
two matrix cells.
}
\examples{
#create an example ID matrix for `IDmat`. Example uses a 100x100 matrix
mat <- matrix(1:10000, nrow = 100)

}
