% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{meltheatmats}
\alias{meltheatmats}
\title{meltheatmats}
\usage{
meltheatmats(mats)
}
\arguments{
\item{mats}{A list of overlap matrices as is produced by `vectorheat`.}
}
\value{
Rturns a list of dataframes in long form, suitable for heat map
plotting via `ggplot2`.
}
\description{
Convenience function to `melt` all the heatmap matirces into a long-form
data.frame format that can be used by `ggplot2`. Retains list naming so that
variables are correctly identified in created data.frames.
}
