% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{plotheatvector}
\alias{plotheatvector}
\title{plotheatvector}
\usage{
plotheatvector(meltlist)
}
\arguments{
\item{meltlist}{A list ofdataframe objects of the format required by
`plotsingleheat`.}
}
\value{
Returns a list of `ggplot2` plot objects
}
\description{
plotheatvector
}
