% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{plotsingleheat}
\alias{plotsingleheat}
\title{plotsingleheat}
\usage{
plotsingleheat(data)
}
\arguments{
\item{data}{}
}
\description{
Function to creat a `ggplot2` heatmap from a single  "melted" matrix of
sampling distribution overlap. Intended to be called by vectorized wrapper function, but could be called on a single element of the list produced by
`meltheatmats` to produce a sinlge heatmap
}
