% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{pref.strength}
\alias{pref.strength}
\title{pref.strength}
\usage{
pref.strength(A.coef, mat, ...)
}
\arguments{
\item{A.coef}{Numeric. Can be a number of vector.  Describes the strength(s) of preference for Habitat A over Habitat B.}

\item{mat}{Matrix.  Tne y matrix produced by create.land.  Converts the Habitat patches of this matrix
(which are labelled) by patch ID to the value(s) of A.coef.}

\item{...}{Additional arguments as necessary.}
}
\value{
A matrix with values of the grown habitat converted to relative strengths of preference OR
an array of matrix values if A.coef is supplied as a vector
}
\description{
Converts the values of the simulated patchy habitat matirx to strength of preference values supplied by the number or vector A.coef.
If A.coef is supplied as vector is vectorizes the process to create a matrix for each value of A.coef
}
