% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habprefmodel.R
\name{prob.sel}
\alias{prob.sel}
\title{prob.sel}
\usage{
prob.sel(mat, ...)
}
\arguments{
\item{mat}{The array of probabilities of selection (output from convert.cell). Anticipates that it will be supplied
as part of muli-model vecotrization.}

\item{...}{Additional arguments as necessary.}
}
\value{
Returns the cell ID for the chosen cell.
}
\description{
Simulate an individual settling the landscape based on habitat preference. The strength of habitat preference
is supplied as part of the generation of the simulated landscape in the term A.coef included in the function
pref.strength. This function only chooses loctions within the matrix based on the supplied parameterization.
}
