% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{probsel}
\alias{probsel}
\title{probsel}
\usage{
probsel(hab.prob, lambda, move.list, coef.d, coef.r, matsize, ...)
}
\arguments{
\item{hab.prob}{matrix, the habitat matrix with cell values representing the
habitat-only probabilities (uncorrected for the mixing proportions with
distance). Typically the second element of the list returned by
simlandscapes.}

\item{lambda}{numeric, the strength of the declining exponental function that controls
how "reluctant" the agent is to move large distances. Must be <1.}

\item{move.list}{list, the current `move.list` being maintained by
`chooseloc`. Used to find the current agent position.}

\item{coef.d}{numeric, the relative weight to give the distance preference.}

\item{coef.r}{numeric, the relative weight to give the habitat preference.}

\item{matsize}{integer, the size in number of cells of one side of the square
matrix on which the simulation is being run.}

\item{...}{}
}
\value{

}
\description{
Makes probabalistic selections of a cell based on the mixture of two
probability matrices. Not intended to be called directly - used by
`chooseloc`
}
\details{
(based on https://scrogster.wordpress.com/2012/04/22/using-r-for-spatial-sampling-with-selection-probabilities-defined-in-a-raster/)
}
