% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullmodel.R
\name{rand.sel}
\alias{rand.sel}
\title{rand.sel}
\usage{
rand.sel(hab.mat, ...)
}
\arguments{
\item{hab.mat}{The matrix containing the "grown" habitats enumerated by the strength of preference
(e.g., the output from pref.strength, not convert.cell.). This matrix is only used to calculate the
size of the landscape from which the agent chooses.}

\item{...}{Additional arguments as necessary.}
}
\value{
Returns a single number: the cell ID of the randomly chosen location.
}
\description{
Simulation of a null model where individuals simply randomly lccate on the landscape with no locational
preferences. Each iteration of this function represents a single individual settling a landscape randomly.
}
