% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conattractionmodel.R
\name{repCon}
\alias{repCon}
\title{repCon}
\usage{
repCon(reps, radius, ID.mat, hab.mat, n.individ, A.coef, ...)
}
\arguments{
\item{reps}{Number of model runs to simulate (each run conisting of `n.individ` number of individuals settling)}

\item{radius}{Sets how many cells away from the occupied cells should be included in the "neighborhood".
Passed to `neighbors`. Passed to `settle.con`.}

\item{ID.mat}{A matrix of the same size as the simulated landscape but with the value of each cell corresponding
to it's cell ID. Passed to `settle.con`.}

\item{hab.mat}{The matrix containing the "grown" habitats enumerated by the strength of preference
(e.g., the output from pref.strength, not convert.cell.) Note that which hab.mat to use (i.e., which
preference strength to use) is arbitrary for the CA model, but the choice must match the `A.coef` supplied (below) andmust
be entered into this function as only the single matrix and not the vector of matirces used to iterate through the HP model.}

\item{n.individ}{The number of individuals included in the ABM.}

\item{A.coef}{The habitat preference parameter that matches the hab.mat supplied.  This is the "key" that
allows the function to interpret each cell of the matrix as either "Habitat A" or "Habitat B".}

\item{...}{Additional arguments as necessary.}
}
\value{
Returns a list, each entry of which is the output of `reps` number of
iterations of `settle.hab` (if vecotrized, for a a single parameterization).
}
\description{
repCon
}
