% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habprefmodel.R
\name{repHab}
\alias{repHab}
\title{repHab}
\usage{
repHab(p.mat, hab.mat, reps, n.individ, A.coef, ...)
}
\arguments{
\item{p.mat}{The probability matrix passed to prob.sel. Typically the output from `convert.cell`.
Passed to `settle.hab`}

\item{hab.mat}{The matrix containing the "grown" habitats enumerated by the strength of preference
(e.g., the output from pref.strength, not convert.cell). Passed to `settle.hab`}

\item{reps}{Number of model runs to simulate (each run conisting of `n.individ` number of individuals settling)}

\item{n.individ}{The number of individuals to simulate for a single model run. Passed to `settle.hab`.}

\item{A.coef}{The habitat preference parameter vector .  This is the "key" that
allows the function to interpret each cell of the matrix as either "Habitat A" or "Habitat B". Passed to settle.hab.}

\item{...}{Additional arguments as necessary.}
}
\value{
Returns a list, each entry of which is the output of `reps` number of
iterations of `settle.hab` (if vectorized, for a a single parameterization).
}
\description{
Generate sampling distributions from settle.hab vectorized across parameterizations.
Allows settle.hab to be iterated reps number of times across multiple parameterizations of A.coef.
}
