% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefood.R
\name{samplesources}
\alias{samplesources}
\title{samplesources}
\usage{
samplesources(
  num_samples,
  food,
  filename = NULL,
  writefile = T,
  returnobject = F
)
}
\arguments{
\item{num_samples}{Integer, the number of samples to colect from each prey
source}

\item{food}{list object of simulated food sources.}

\item{filename}{name of file to save. Passed to `savesources`.}

\item{writefile}{boolean, whether to write the csv to file needed by MixSIAR
default is TRUE. Passed to `savesources`.}

\item{returnobject}{boolean, whether to return an object within the current
environment, default is FALSE. Passed to `savesources`.}
}
\value{
Writes a .csv to file containing the sample-based estimates for prey
source isotop distribution in `MixSIAR` format.
}
\description{
Function to simulate sample-based estimate of prey source isotopic values for
subsequent mixture decomposition.
}
