% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefood.R
\name{samplesourcesvector}
\alias{samplesourcesvector}
\title{samplesourcesvector}
\usage{
samplesourcesvector(n_vec, food, filepath = NULL, writefile = T,
  returnobject = F, ...)
}
\arguments{
\item{n_vec}{numeric, vector of sample sizes for sampling from diet sources.}

\item{food}{list object of simulated food sources.}

\item{writefile}{boolean, whether to write the csv to file needed by MixSIAR
default is TRUE. Passed to `savesources`.}

\item{returnobject}{boolean, whether to return an object within the current
environment, default is FALSE. Passed to `savesources`}

\item{filename}{name of file to save. Passed to `savesources`.}
}
\value{
Saves diest source sampled isotope distributions to files of format
used by `MixSIAR`
}
\description{
Function to simulate sampling of potential diet source populations across a
set of values for the size of the sample.  Largely a convenience function to
run `samplesources` across multiple values of sample size.
}
