% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefood.R
\name{savediscrimination}
\alias{savediscrimination}
\title{savediscrimination}
\usage{
savediscrimination(food, filename = NULL, discr_carb = c(0, 0, 0),
  discr_nit = c(0, 0, 0), writefile = T, returnobject = F)
}
\arguments{
\item{food}{list object of simulated food sources}

\item{filename}{name of file to save}

\item{discr_carb}{vector of carbon discrimination values,
one per diet source.  Default is no discrimintation}

\item{discr_nit}{vector of nitrogen discrimination values,
one per diet source.  Default is no discrimintation}

\item{writefile}{boolean, whether to write the csv to file needed by MixSIAR
default is TRUE}

\item{returnobject}{boolean, whether to return an object within the current
environment, default is FALSE}
}
\value{
either writes a MixSIAR formatted csv to file or a dataframe object
}
\description{
function to simulate isotope discrimination data in the format needed by

MixSIAR. Defaults to no discrimination factors.
}
