% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefood.R
\name{savesources}
\alias{savesources}
\title{savesources}
\usage{
savesources(
  food,
  filename = NULL,
  mu_carb,
  sd_carb,
  mu_nit,
  sd_nit,
  popsize,
  writefile = T,
  returnobject = F
)
}
\arguments{
\item{food}{list object of simulated food sources}

\item{filename}{name of file to save}

\item{mu_carb}{vector whose length = 'num_sources', each entry of which is
the mean 13C value of the diet source population to simulate}

\item{sd_carb}{vector whose length = 'num_sources', each entry of which is
the standard deviation of the 13C of the diet source population to simulate
#' @param mu_nit vector whose length = 'num_sources', each entry of which is
the mean 15N value of the diet source population to simulate}

\item{sd_nit}{vector whose length = 'num_sources', each entry of which is
the standard deviation of the 15N of the diet source population to simulate}

\item{popsize}{vector whose length = 'num_sources', each entry of which is
the size of the diet source population to simulate}

\item{writefile}{boolean, whether to write the csv to file needed by MixSIAR
default is TRUE}

\item{returnobject}{boolean, whether to return an object within the current
environment, default is FALSE}
}
\value{
either writes a MixSIAR formatted csv to file or a dataframe object
}
\description{
Convenience function to save simulated prey source isotope parameters in
format that can be used by `MixSIAR`.  Can either write the data to a .csv
file and/or return as a dataframe object. Can be called independently if
mean and standard deviation of the prey source isotope distributions are
known (e.g., if simulating that prey source isotope disrtibutions are known
perfectly) or can be called from `samplesources` to save simulated sampling
of prey source isotope values.
}
