% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habprefmodel.R
\name{settle.hab}
\alias{settle.hab}
\title{settle.hab}
\usage{
settle.hab(n.individ, p.mat, hab.mat, A.coef, ...)
}
\arguments{
\item{n.individ}{The number of individuals the model should simulate.
Repeats prob.sel this number of times for a single model run.}

\item{p.mat}{The probability matrix passed to prob.sel. Typically the output from convert.cell}

\item{hab.mat}{The matrix containing the "grown" habitats enumerated by the strength of preference
(e.g., the output from pref.strength, not convert.cell).}

\item{A.coef}{The habitat preference parameter that matches the hab.mat supplied.
This is the "key" that allows the function to interpret each cell of the matrix as either "Habitat A" or "Habitat B".}

\item{...}{Additional arguments as necessary.}
}
\value{
A list, the first element of which is the set of coordinates for the settled location of each individual and
the 2nd element of which is the proportion of those indviduals that settled within "Habitat A".  If the model is iterated,
returns a list of such lists.
}
\description{
Repeates prob.sel for n.individ number of individuals. Completes a single model run.
}
