% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullmodel.R
\name{settle.rand}
\alias{settle.rand}
\title{settle.rand}
\usage{
settle.rand(n.individ, hab.mat, A.coef, ...)
}
\arguments{
\item{n.individ}{The number of individuals to simulate for a single model run.}

\item{hab.mat}{The matrix containing the "grown" habitats enumerated by the strength of preference
(e.g., the output from pref.strength, not convert.cell.) Note that which hab.mat to use (i.e., which
preference strength to use) is arbitrary for the null model, but the choice must match the A.coef supplied (below).}

\item{A.coef}{The habitat preference parameter that matches the hab.mat supplied.  This is the "key" that
allows the function to interpret each cell of the matrix as either "Habitat A" or "Habitat B".}

\item{...}{Additional arguments as necessary.}
}
\value{
A list, the first element of which is the set of coordinates for the settled location of each individual and
the 2nd element of which is the proportion of those indviduals that settled within "Habitat A".  If the model is iterated,
returns a list of such lists.
}
\description{
A convenience function that allows 'rand.sel' to be iterated across multiple individuals ('n.individ')
Allows the entire ABM to be run an arbitray number of times  (see `reps` below) in order to generate a
sampling distribution of apparent selections generated by the model.
}
