% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{simlandscapes}
\alias{simlandscapes}
\title{simlandscapes}
\usage{
simlandscapes(A.coef, matsize, count.max = 200, n.clusters,
  size.clusters)
}
\arguments{
\item{A.coef}{Numeric. Can be a number of vector.  Describes the strength(s)
of preference for Habitat A over Habitat B.}

\item{matsize}{integer, the size of a side of the suare matrix to be created.}

\item{count.max}{interger, the maximum number of iterations throught the
while-loop.  Default value is 200}

\item{n.clusters}{integer, the number of indpendent habitat patches of
"type A" to grow.}

\item{size.clusters}{integer, the target average size (in number of matrix
cells) for each patch.}
}
\value{
A list each element of which contains the "habitat matrix" and
"preference strength" matrix for the given value of A.coef. There is an
element in the list for each value of A.coef supplied.
}
\description{
simlandscapes
}
