% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{simpatches}
\alias{simpatches}
\title{simpatches}
\usage{
simpatches(matsize, count.max = 200, n.clusters, size.clusters)
}
\arguments{
\item{matsize}{integer, the size of a side of the suare matrix to be created.}

\item{count.max}{interger, the maximum number of iterations throught the
while-loop.  Default value is 200}

\item{n.clusters}{integer, the number of indpendent habitat patches of
"type A" to grow.}

\item{size.clusters}{integer, the target average size (in number of matrix
cells) for each patch.}
}
\value{
Returns a matrix of size `matsize` x `matsize` containing
approximately `n.clusters` patches of approximately `size.clusters` size.
}
\description{
Function to simulate a patchy habitat matrix by "growing" one habitat type
from a cellular automota-like loop.
}
