% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{singleheatmat}
\alias{singleheatmat}
\title{singleheatmat}
\usage{
singleheatmat(data1, data2)
}
\arguments{
\item{data1}{The first data set to compare. Anticipates the output of one
of the `rep...` calls (e.g., `repRand`).}

\item{data2}{The second data set to compare. Anticipates the output of one
of the `rep...` calls (e.g., `repRand`).}

\item{param1}{The numeric parameter object originally fed to the simulation
model that generated data set 1.}

\item{param2}{The numeric parameter object originally fed to the simulation
model that generated data set 2.}
}
\description{
Function to calculate matrix of overlap between one distribution and the
range of another. Builds the matrix upon which a heatmap can act.
}
