% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{startmat}
\alias{startmat}
\title{startmat}
\usage{
startmat(matsize)
}
\arguments{
\item{matsize}{integer, the size of a side of the suare matrix to be created.}
}
\value{
a square matrix whose sides are of size `matsize` and all cell values
 are "1".
}
\description{
Creates a square matirx of supplied size with all cell values being 1.  This
is the matrix anticipated by the habitat growth functions.
}
