% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{vectorheat}
\alias{vectorheat}
\title{vectorheat}
\usage{
vectorheat(mod1, mod2)
}
\arguments{
\item{mod1}{A list of model objects, each element of which will be the first
data set to compare in the `singleheatmat` function. This function is
intended to recieve the first element of the list produced by `getcombos`.}

\item{mod2}{The second list of model objects, each element of which will be
the second data set to compare in the `singleheatmat` function. This
function is intended to recieve the second element of the list produced by
 `getcombos`.}
}
\value{
Returns a list, each element of which is matrix of the sampling
distribution overlaps as produced by `singleheatmat`.
}
\description{
The vectorized wrapper function for `singleheatmat` that can walk through the
paired list of model combiantions produced by `getcombos` and produce matrices
of overlap data with which heat maps can be generated.
}
\examples{
modelcomb <- getcombos(X = list("NULL"=NULLmod, "HP"=HABmod,
"CA"=CAmod), Y = list("NULL"=NULLmod, "HP"=HABmod, "CA"=CAmod))
 heats <- vectorheat(modelcomb[[1]], modelcomb[[2]])
}
