% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{matrixPythagoras}
\alias{matrixPythagoras}
\title{matrixPythagoras}
\usage{
matrixPythagoras(position1, position2, matsize)
}
\arguments{
\item{position1}{integer, first positions in distance calc - must be supplied
as cell ID}

\item{position2}{integer, second positions in distance calc - must be
supplied as cell ID
#' @param matsize integer, the size in number of cells of one side of the square
matrix on which the simulation is being run.}
}
\value{
returns the pythogorean distance between two coordinates (supplied as
matirx indices)
}
\description{
Uses the Pythagorean theorem to calculate the strightline distance between
two matrix cells.
}
\examples{
#create an example ID matrix for `IDmat`. Example uses a 100x100 matrix
mat <- matrix(1:10000, nrow = 100)

}
