% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_and_visualization.R
\name{plotsingleheat}
\alias{plotsingleheat}
\title{plotsingleheat}
\usage{
plotsingleheat(data)
}
\arguments{
\item{data}{A single long-form dataframe object containing the
parameterizations of the models to be compared and the calculated pairwise
 unidirectional overlap in distributions.}
}
\value{

}
\description{
Function to creat a `ggplot2` heatmap from a single  "melted" matrix of
sampling distribution overlap. Intended to be called by vectorized wrapper function, but could be called on a single element of the list produced by
`meltheatmats` to produce a single heatmap
}
